requirements.txt
altgraph==0.17.4
iniconfig==2.0.0
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2025.0
pytest==8.3.4
pytube==15.0.0
setuptools==75.8.0
